generalize the panel_ids function so it can be included in local_tools.py and make corresponding changes in Tanzania
